[SYSTEMDS-113] New builtin functions for linear regression algorithms
This patch adds 4 new DML-bodied builtin functions:

  * lm
  * lmDS
  * lmCG
  * lmpredict

These functions are to be used for linear regression.
`lm` calls either `lmDS` or `lmCG` depending on the size of the input
matrices. `lmpredict` can be used
to predict the class of a feature vector with the weights usually
calculated before by use of `lm`.

This also adds a file used for documentation of builtin functions, which
currently only documents those new added builtins.

Beware that lm currently is non-functional and the test cases will fail,
because of a namespace conflict when using multiple different DML-bodied
builtin functions.

Closes #14.