Add multi-threaded kahan sum aggregation. Optimize aggregation for different datatypes.
This commit improves the sum operation to use multiple threads. This is as fast as the current matrix implementation for all datatypes except String, which is slower and it should be discussed how a `sum` on Strings should even work, and Boolean, where it is a lot faster. If the matrix is empty though it optimizes it to be a metadata operation (with non zero count), while the tensor does not yet support this.

Reshape instruction should also be as fast as matrix reshape and makes use of copy on write.

Note that all instructions are only implemented for CP-mode.