Tensor Constructor, Sum and toString built-ins for DML
This commit adds the built-in `tensor()` and support for tensors for the
built-ins `sum()` and `toString()`. The implementation is not very clean
and there are a lot of TODOs (which should all be marked by comments).

Note that the CostEstimatorStaticRuntime will probably not be possible to
accurately work on `tensor()` if we want to support all functionality of
`tensor()` currently implemented.

The `sum()` implementation is also not very in-depth and only very basic.

`toString()` works nicely, although we might have to think about how we
want to truncate the output if the size is too large. At the moment it
checks for the first two dimensions, but this of course does not really
limit the output size.

I was not sure at which point I should switch from a modifications license
to normal license by TU Graz, because I sometimes copied parts of already
implemented features.