Add support for most datatypes on `tensor()`
This additionally fixes a bug in `DenseBlockBool` where the copying of another DenseBlock did not function correctly.

For all datatypes except `String` it is now supported to fill the tensor with it, while the tensor keeps the datatype. Also adds print functionality for all `tensor` containing non double data.

Note that the `String` functionality could not be implemented at the moment, since we allow data to be passed as list of values in a `String`, and therefore we can not infer if the user wants to fill the tensor with the `String` as value or with data in the `String` concatenated by spaces.

If we pass a `tensor` as an `data`-argument to the constructor, the new tensor keeps the same datatype.