Made code more compact
This commit tries to simplify the code a bit, this leads to some differences compared to the matrix sum aggregation.

I did some performance tests for a specialized sum implementation for FP64 (making use of `DenseBlock.valuesAt()`), but got worse performance for a variety of block sizes (100kx1k, 10kx1k, 1kx1k, 100x1k and 10x1k), so I will keep the general implementation for now and we can have another look into it later.