Add tests, binary operations and federated matrices
We can create matrices from files with `.mtd` or numpy. Basic arithmetic is implemented and tested, as well as some aggregations.

Federated works, but I have not added tests because we have to start a worker and we might want to discuss how we distribute the systemds jar. Currently either a standalone jar or our SystemDS jar + classes and libs has to placed in the directory `systemds-java` for the implementation to work, or the java jmlc gateway is manually started and the method `set_aut_start_jmlc(False)` is called.

I did not yet have a detailed look into packaging and releasing on pip, but it seems easy with `setuputils`.

Currently on top of todo list are test cases for federated.

For comparisons like `<`, `>=` etc. we also have to talk about either giving a `DAGNode` back on which `.compute()` has to be called, or calling `.compute()` for both parts and comparing the results. Further discussions which actions except `.compute()` actually start the execution in SystemDS is necessary.